feat : 폼 로그인 구현, DoctorQna CRUD 로그인 연동, Entity 수정으로 인한 전체적인 코드 리팩토링

+++ src/main/java/com/codelion/animalcare/domain/doctorqna/controller/AnswerController.java
@@ -6,14 +6,18 @@ import com.codelion.animalcare.domain.doctorqna.dto.request.AnswerUpdateRequestD
 import com.codelion.animalcare.domain.doctorqna.service.AnswerService;
 import com.codelion.animalcare.domain.doctorqna.service.QuestionService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.server.ResponseStatusException;
 
 import javax.validation.Valid;
+import java.security.Principal;
 
 @RequiredArgsConstructor
 @Controller
@@ -23,41 +27,60 @@ public class AnswerController {
     private final QuestionService questionService;
 
     //답변 작성
+    @PreAuthorize("isAuthenticated()")
     @PostMapping("/usr/doctor-qna/{questionId}/answers/write")
-    public String save(Model model, @PathVariable Long questionId, @Valid AnswerSaveRequestDto answerSaveRequestDto, BindingResult bindingResult){
+    public String save(Model model, @PathVariable Long questionId, @Valid AnswerSaveRequestDto answerSaveRequestDto, BindingResult bindingResult, Principal principal){
 
         if(bindingResult.hasErrors()) {
             model.addAttribute("question", questionService.findById(questionId));
             return "doctorqna/doctorQnaDetail";
         }
 
-        answerService.save(questionId, answerSaveRequestDto);
+//        if(!answerService.isDoctor(principal)) {
+//            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "의사만 답변을 작성할 수 있습니다.");
+//        }
+
+        answerService.save(questionId, answerSaveRequestDto, principal);
         return "redirect:/usr/doctor-qna/%d".formatted(questionId);
     }
 
-    //TODO : 아래 기능들은 로그인 후 구현?
-
+    @PreAuthorize("isAuthenticated()")
     @GetMapping("/usr/doctor-qna/{questionId}/answers/{answerId}/modify")
-    public String modify(Model model, @PathVariable Long questionId, @PathVariable Long answerId, AnswerUpdateRequestDto answerUpdateRequestDto){
+    public String modify(Model model, @PathVariable Long questionId, @PathVariable Long answerId, AnswerUpdateRequestDto answerUpdateRequestDto, Principal principal){
+
+        if(answerService.answerAuthorized(answerId, principal)){
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "수정권한이 없습니다.");
+        }
 
         model.addAttribute("answer", answerService.findById(answerId));
 
         return "/doctorqna/doctorQnaAnswerModifyForm";
     }
-
+    @PreAuthorize("isAuthenticated()")
     @PostMapping("/usr/doctor-qna/{questionId}/answers/{answerId}/modify")
-    public String modify(@PathVariable Long questionId, @PathVariable Long answerId, @Valid AnswerUpdateRequestDto answerUpdateRequestDto, BindingResult bindingResult){
+    public String modify(@PathVariable Long questionId, @PathVariable Long answerId, @Valid AnswerUpdateRequestDto answerUpdateRequestDto, BindingResult bindingResult, Principal principal){
 
         if(bindingResult.hasErrors()) {
             return "/doctorqna/doctorQnaAnswerModifyForm";
         }
+
+        if(answerService.answerAuthorized(answerId, principal)){
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "수정권한이 없습니다.");
+        }
+
         answerService.update(questionId, answerId, answerUpdateRequestDto);
         return "redirect:/usr/doctor-qna/%d".formatted(questionId);
     }
 
     //답변 삭제
+    @PreAuthorize("isAuthenticated()")
     @GetMapping("/usr/doctor-qna/{questionId}/answers/{answerId}/delete")
-    public String delete(@PathVariable Long questionId, @PathVariable Long answerId) {
+    public String delete(@PathVariable Long questionId, @PathVariable Long answerId, Principal principal) {
+
+        if(answerService.answerAuthorized(answerId, principal)){
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "수정권한이 없습니다.");
+        }
+
         answerService.delete(questionId, answerId);
         return "redirect:/usr/doctor-qna/%d".formatted(questionId);
     }
